Move the entire capstone-deploy repository to deploy/
We didn't have a strong reason to create capstone-deploy as a second
repository, other than it tends to be a best practice to separate the
two. Do we have any reason to follow that best practice?

With everything in one repository, the project and the code to deploy it
stay in lockstep with each other from development, through CI, to
delivery.

This is a direct copy of everything in capstone-deploy to a new deploy/
directory in capstone. Literally:

    $ mkdir capstone/deploy/
    $ cd capstone-deploy
    $ git checkout 8843026e35a16f333c52addc88c171aeab1a3957
    $ cp -R . ../capstone/deploy/

CI, etc, are not updated.

Reminds me of: http://danluu.com/monorepo/

Change-Id: I331ac5ca88f13db67af979725d45117b5463aac8