Fixes logging the incorrect domain
The code was logging the user's domain that was retrieved from the 2.0
API and saying the user is not in it. This is clearly incorrect since
the 2.0 API is telling us that it is the user's domain. Instead we want
to log the domain the user claims to be in.

Change-Id: Idc57301013df9d4a269609b4cb2e561daae18380