Add ability to deploy capstone from venv
This commit introduces inventory variables that allow us to deploy capstone and
all it's dependencies as a venv. This is nice because it makes it so the only
python packages needed to install capstone are virtualenv-tools and httplib2.

This commit also changes the capstone version variable to match how venvs are
built/named by the Jenkins build job. It also sets an ansible configration
option to make hostnames really short. This was something that needed to be
done in order to deploy against hosts with really long hostnames.

Change-Id: Ic6a94033e6136b11a84864d9a8bc83a71313595a