Do not use raise_for_status() in tests
It's a simple shortcut, but it allows any non-2xx code to pass the test,
which is too relaxed. When a test does fail, the test log contains
nothing more than the actual status code, ignoring the rest of the HTTP
response that keystone/capstone returned. By passing the specific status
code to assertEqual and using the third argument of assertEqual(), we
get the complete response body which makes it easier to debug failures.

Change-Id: I75ab308ad4cfd86e342937fd46ce1238453701fe