Eliminate dedicated coverage job
Unlike upstream, we have no reason to run coverage separately, so this
merges the coverage commands into the py27 job. Coverage is now always
reported with the py27 test results. This means we don't have to create
two environments with the exact same dependencies just to run a
different command in each.

Change-Id: If160fe400ef4af1481874528f993920b243155c3