Make venv_vars.yml explicit
This var file is used to set variables that are needed to run against
non-developer environments. These variables are also really well documented
inline and it's assumed that it will be used in conjunction with an inventory
file that specifies all the secret bits in order to run successfully.

Also, by not assuming defaults it makes it easier to diagnose issues with
inventory when the play fails immediately instead of doing something unexpected
because a default was set.

Change-Id: I4561e5e8991ceaf5b0c7bc768846d927333948a1