Sort of fix the dev pipeline
The dev pipeline was completely broken, so this is an attempt to make it
useful again. The working directory values appeared in the wrong place,
and the environment variables were not being passed to the
internal/shell step as intended.

After running wercker dev --wercker-yml="wercker-deploy.yml" now, you're
left trailing all the logs of a running deployment. You can CTRL+C to
start working inside the container if desired.

Change-Id: Ibf474182442146ae82ad707e7070b5a6c3f7b856