Remove wercker-deploy.yml
The wercker-deploy.yml file didn't contain anything not already tested by the
wercker-integration.yml pipeline. The only thing specific to the
wercker-deploy.yml file was a dev pipeline. We can consolidate the number of
wercker files in the project by removing wercker-deploy.yml and if we really
need a dev pipeline, we can define one in the wercker-integration.yml file.

This commit just removes unnecessary CI cruft from the project.

Change-Id: I08363a0cb414cdb457cf852b19b2704a1d47933d