Give polling_period a default in capstone
The polling_period configuration option didn't have a default in capstone.
Instead, it was being deployed with a default value by the deployment tooling.
This commit makes it so that polling_period has a default value in capstone.
This commit also changes the type from cfg.StrOpt to cfg.IntOpt and removes
casting the option to an int() in code.

Change-Id: I8b5a2916affabb450c0165f2f27fea273d67ef65