Separate testing variables into their own file
Previously, we were just stuffing all variables needed to run the deployment
and integration tests in the inventory file. This made it hard to distinguish
which variables in the inventory were used for testing and which were used for
real deployments.

This commit separates the testing variables into their own file. What is in the
inventory file should only be what is needed to deploy capstone to production
or staging.

Change-Id: I23da798f02e2986641c7f36b21d60a453d038d04