Remove unused variables from inventory
These variables used to be referenced in the plays until they were refactored
out into a file specifically for staging and production deployments. We will
always be referencing them there and passing that file to ansible on the
command line. This makes the inventory file and the handling of deploying
locally or not less complicated.

Change-Id: I225d898b0c45690b7a85b4f1ad6bdf10e52e6f16