Add testing clarification to README.rst
In the documentation we made it pretty clear that the integration tests needed
to be invoked separately if you wanted to run them. With the mock_v2 service
we can run the integration tests on the same host as capstone. This commit
updates the docs to reflect that change since the integration environment/tests
are no longer specifically excluded from the tox run.

Change-Id: I297de07ec20f6677328e16bf04aad89c70f9a303