Restore py27 job
I removed the envlist thinking that py27 would be automatically run
along with lint and integration, but instead we've *only* been running
lint and integration automatically in the gate.

The order of the envlist also determines the order of execution, so the
lint job will fail on syntax errors before the integration job fails
with 500's, for example.

Change-Id: I4b06fa8816ff4e11c31ea798c8820c5c7ee5c8ed