Remove admin_project_id
This was something that was in config as a side-effect of using the service
user as the integration test user. As far as capstone is concerned, it doesn't
need the service users project ID in order to do the things it needs to do. We
have been running Capstone without a service user project ID, but the
configuration option was still being set in capstone/conf.py.

This commit removes conf.admin_project_id.

Change-Id: I787e3f43121f6aa68dca02597a40e063edf04ae2