Document unsupported use cases appropriately
This list has caused a bit of confusion because people have assumed that
Capstone supported all of these use cases today, and then were confused
when Capstone did not respond to the corresponding APIs as expected. A
bit of additional italicized text will make the state of these use cases
explicit until they are implemented.

Change-Id: Ia7c78d3ce915be45dcff35f18cace69a34f4bb12